+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# History of changes
+
 ## [1.0.0] - 2020-05-23
 
 The firs release include 6 field types (with many parameters), 2 input actions, submit button, reset Button.
@@ -5,3 +7,8 @@ The firs release include 6 field types (with many parameters), 2 input actions, 
 ## [1.0.1] - 2020-05-26
 
 Added the example of usage with 2 fields.
+
+## [1.0.2] - 2020-11-05
+
+- Added new option to manage the validation, reset and get the response with external buttons.
+- Added example with new functionallity.